pyproject.toml

### pyproject.toml

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "router-ids"
version = "0.1.0"
description = "Lightweight intrusion detection system for Raspberry Pi routers"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

dependencies = [
    "scapy>=2.5.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
]

[project.urls]
Homepage = "https://github.com/chetan-914/router-ids"
Repository = "https://github.com/chetan-914/router-ids.git"

[tool.setuptools]
packages = ["router_ids"]

[tool.setuptools.package-data]
router_ids = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false